顔計算機能完成
