+++ Bemayu.xcodeproj/xcuserdata/cmstudent.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,116 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "303C9503-EB7E-45A9-B85E-E37196984E3B"
+   type = "1"
+   version = "2.0">
+   <Breakpoints>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700"
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "Bemayu/FaceScan/Camera/FaceScan.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "14"
+            endingLineNumber = "14"
+            landmarkName = "LiveFeedViewController"
+            landmarkType = "3">
+            <Locations>
+               <Location
+                  uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700 - 48f8f0c68952c252"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "Bemayu.LiveFeedViewController.previewLayer.getter : __C.AVCaptureVideoPreviewLayer"
+                  moduleName = "Bemayu"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmstudent/Documents/%E5%8D%92%E5%88%B6/Bemayu/Bemayu/FaceScan/Camera/FaceScan.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "14"
+                  endingLineNumber = "14"
+                  offsetFromSymbolStart = "168">
+               </Location>
+               <Location
+                  uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700 - 85f2ebcbc67c9aa6"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "Bemayu.LiveFeedViewController.previewLayer.setter : __C.AVCaptureVideoPreviewLayer"
+                  moduleName = "Bemayu"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmstudent/Documents/%E5%8D%92%E5%88%B6/Bemayu/Bemayu/FaceScan/Camera/FaceScan.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "14"
+                  endingLineNumber = "14"
+                  offsetFromSymbolStart = "108">
+               </Location>
+               <Location
+                  uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700 - b00698e8d4706457"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "Bemayu.LiveFeedViewController.previewLayer.modify : __C.AVCaptureVideoPreviewLayer"
+                  moduleName = "Bemayu"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmstudent/Documents/%E5%8D%92%E5%88%B6/Bemayu/Bemayu/FaceScan/Camera/FaceScan.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "14"
+                  endingLineNumber = "14"
+                  offsetFromSymbolStart = "0">
+               </Location>
+               <Location
+                  uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700 - 9a2e320678de31a8"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "Bemayu.LiveFeedViewController.previewLayer.modify : __C.AVCaptureVideoPreviewLayer with unmangled suffix &quot;.resume.0&quot;"
+                  moduleName = "Bemayu"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmstudent/Documents/%E5%8D%92%E5%88%B6/Bemayu/Bemayu/FaceScan/Camera/FaceScan.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "14"
+                  endingLineNumber = "14"
+                  offsetFromSymbolStart = "20">
+               </Location>
+               <Location
+                  uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700 - 46a22daa1462b1ac"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "Bemayu.LiveFeedViewController.init(frame: __C.CGRect) -&gt; Bemayu.LiveFeedViewController"
+                  moduleName = "Bemayu"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmstudent/Documents/%E5%8D%92%E5%88%B6/Bemayu/Bemayu/FaceScan/Camera/FaceScan.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "14"
+                  endingLineNumber = "14"
+                  offsetFromSymbolStart = "176">
+               </Location>
+               <Location
+                  uuid = "8F2E7E18-DDC2-4DDC-8F99-37E1E4B53700 - 50228cf7b6dd5544"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "Bemayu.LiveFeedViewController.init(coder: __C.NSCoder) -&gt; Swift.Optional&lt;Bemayu.LiveFeedViewController&gt;"
+                  moduleName = "Bemayu"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmstudent/Documents/%E5%8D%92%E5%88%B6/Bemayu/Bemayu/FaceScan/Camera/FaceScan.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "14"
+                  endingLineNumber = "14"
+                  offsetFromSymbolStart = "144">
+               </Location>
+            </Locations>
+         </BreakpointContent>
+      </BreakpointProxy>
+   </Breakpoints>
+</Bucket>
