--- Bemayu/FaceScan/Camera/Test.swift
@@ -1,14 +1,0 @@
-//
-//  Test.swift
-//  Bemayu
-//
-//  Created by cmStudent on 2022/11/15.
-//
-
-import AVFoundation
-import UIKit
-import Vision
-
-class Test: UIViewController {
-    
-}
